requirements.txt
marshmallow
smart-open==2.2.1
requests
aws-secretsmanager-caching
pyyaml==5.4.1
gretel-client
